Cleaned up some code, named some variables for better understanding, added option for building entire mx library if desired.

+++ src/SimplifiedNote.h
@@ -1,14 +1,14 @@
 #pragma once
 
+#include <string>
 #include <iostream>
 #include "NoteEnums.h"
 #include "mx/api/ScoreData.h"
 
 using namespace noteenums;
 
-
-//Simpler way to represent notes compared to mx. Used to make it easier to make additions
-//in the future.
+//Simpler way to represent notes compared to mx. Used to make it easier to make 
+//additions in the future.
 class SimplifiedNote {
 	private:
 		const Note note;
@@ -17,16 +17,19 @@ class SimplifiedNote {
 		SimplifiedNote() : note(Note::C_4), duration(Duration::Whole) {}
 		SimplifiedNote(mx::api::NoteData);
 		SimplifiedNote(Note n, Duration d);
+
 		Note getNote();
 		Duration getDuration();
+
 		bool operator < (const SimplifiedNote &rhs) const {
 			return (static_cast<int>(note) < static_cast<int>(rhs.note));
 		}
 		bool operator > (const SimplifiedNote &rhs) const {
 			return (static_cast<int>(note) > static_cast<int>(rhs.note));
 		}
-		friend inline std::ostream& operator<< (std::ostream& out, const SimplifiedNote n) {
-			out << static_cast<int>(n.note);
+		friend inline std::ostream& operator<< (std::ostream& out, 
+							const SimplifiedNote n) {
+			out << n.note;
 			return out;	
 		};
 };
